Create TestNotes
Project is here: https://github.com/users/bodnerry/projects/1
There is no code at this moment. Motivated by a contest entry, The lack of code kills ot for the contest, but the project is on-going. ADHD can affect a person's ability to pull a project together sometimes. However, The only code would be very minimal in order to learn stuff, as the idea really needs more user research to figure what should be in the code. That said, the design is progressing and I'm planning to code up some basic linking to activity-tracking and pushing notifications to the remote device.